comments, suggestions and expansions
I've added some comments and suggestions on liams code mostly. Some suggestions for william too. I ended up changing some archaic variable names and putting more things under exports etc. All thats covered in comments on top of each script i did anything to. The main change is to our filesystem. i organised a lot of it and made a reusable file for any system that may be usable on another body. All these changes can be reverted or ignored since its on an alternative branch but if you all could take a look and if you like it commit what your okay with to the main branch that would be much appreciated.